requirements.txt
streamlit
langchain
yagmail
keyring
pydantic
pymongo
openai
pytz
python-dotenv